Roll tools.py into main GUI
