+++ .gitignore
@@ -13,3 +13,4 @@
 !lib/gui/.cache/icons
 !scripts
 !plugins
+!tools
