+++ lib/gui/__init__.py
@@ -1,9 +1,7 @@
-from lib.gui.console import ConsoleOut
 from lib.gui.command import CommandNotebook
 from lib.gui.display import DisplayNotebook
-from lib.gui.settings import Config
 from lib.gui.stats import CurrentSession, SavedSessions
 from lib.gui.statusbar import StatusBar
 from lib.gui.tooltip import Tooltip
-from lib.gui.utils import Images
+from lib.gui.utils import Config, ConsoleOut, Images, Options
 from lib.gui.wrapper import ProcessWrapper
