+++ tools.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
+""" The master tools.py script """
 import sys
 # Importing the various tools
-from tools.sort import SortProcessor
-from tools.effmpeg import Effmpeg
-import lib.cli as cli
+import tools.cli as cli
+from lib.cli import FullHelpArgumentParser, GuiArgs
 
 # Python version check
 if sys.version_info[0] < 3:
@@ -13,6 +13,7 @@ if sys.version_info[0] == 3 and sys.version_info[1] < 2:
 
 
 def bad_args(args):
+    """ Print help on bad arguments """
     PARSER.print_help()
     exit(0)
 
@@ -23,18 +24,17 @@ if __name__ == "__main__":
     _tools_warning += "understand how it works."
     print(_tools_warning)
 
-    PARSER = cli.FullHelpArgumentParser()
+    PARSER = FullHelpArgumentParser()
     SUBPARSER = PARSER.add_subparsers()
-    EFFMPEG = Effmpeg(
-            SUBPARSER, "effmpeg",
-            "This command allows you to easily execute common ffmpeg tasks.")
-    SORT = SortProcessor(
-            SUBPARSER, "sort",
-            "This command lets you sort images using various methods.")
-    GUIPARSERS = {'effmpeg': EFFMPEG, 'sort': SORT}
-    GUI = cli.GuiArgs(
-            SUBPARSER, "gui",
-            "Launch the Faceswap Tools Graphical User Interface.", GUIPARSERS)
+    EFFMPEG = cli.EffmpegArgs(SUBPARSER,
+                              "effmpeg",
+                              "This command allows you to easily execute common ffmpeg tasks.")
+    SORT = cli.SortArgs(SUBPARSER,
+                        "sort",
+                        "This command lets you sort images using various methods.")
+    GUI = GuiArgs(SUBPARSER,
+                  "gui",
+                  "Launch the Faceswap Tools Graphical User Interface.")
     PARSER.set_defaults(func=bad_args)
     ARGUMENTS = PARSER.parse_args()
     ARGUMENTS.func(ARGUMENTS)
